Add KeePass CLI (xclip-compatible) program
Lately I've been wondering how to organize the scripts/utilities that I use. The
original thought was to create a separation between my dotfiles and those
scripts between two different repositories:

 - `dots` (this one) would maintain the dotfliles
 - `belt` would maintain the scripts/utilities

The thing is... with GNU Stow, the process of "installing" the dotfiles became so
simple that I've been tempted to do the same thing with my scripts. With
`bootstrap.sh` in progress, the usage of all these configurations and
customizations would be blazingly easy and fast.

kpcli (KeePass Command Line Interface) is a program written in Perl
by Lester Hightower. <http://kpcli.sourceforge.net>

This version, in particular, is a fork of the original kpcli but
intended to work with `xclip`. <https://github.com/colobas/kpcli-xclip>

I do not hold any form of authorship on those programs.