Add Random Wal program
From the README:

Random Wal (rwal in short) is a simple and straightforward script that randomly
selects one of the available wallpapers within your computer and uses pywal to
generate a beautiful colourscheme for your terminal-based applications and other
programs.

This is (still) a very raw implementation of one simple thing that I've wanted
to automate in my system: the generation and selection of "themes" based on my
favourite wallpapers.